R7
-- Set botania back to default textures; grouped the alternate ones with
Unity
-- fixed cofhcore json for nether aluminum (generates correctly now)
-- Updated default keybinds once again, with both maps supported. If you
want it, go into the minecraft keybind options and select "reset keys."
This should not override your personal key configuration, if it does
that is a bug.
-- Enabled vanilla enderchest completely
-- Removed some books from the bag o' books that were causing issues,
the book bag no longer lights you on fire when you open it. For the time
being it contains a lot less books unfortunately.
-- Fixed tcon manasteel being the wrong mining level
-- added shapeless recipe to convert enhanced portals diamond nugget to
the translocators one
-- Enabled the culinary generator, the recipe is balanced almost
directly against a botania Gourmaryllis mana --> rf setup. This is for
people who want that setup in a single block.
-- Added recipe for reliquary eye of the storm, because charged creepers
are a pain in the ass to get.
-- Added vanilla recipe for knowledge fragments, to make obtaining
research points easier for those who want to take a bit less punishing
and grindy path through thaumcraft.
-- Added botania recipe for knowledge fragment by infusing a book with
mana. Information about this is added to the lexicon under "Thaumcraft
Integration"
-- Completely replaced chiseltones with a modtweaker script I made
myself. Should be fairly easy to understand. Also changed the recipe for
the base ztones block to something that makes more sense considering you
will be able to chisel every single ztones block from it.